add classes

+++ BL/SharedModels/CartModel.cs
@@ -10,7 +10,7 @@ namespace BL.SharedModels
     {
         public List<CartListModel> CartItem { get; set; }
         public int UserID { get; set; }
-        public int OrderTotalPrice { get; set; }
+        public double OrderTotalPrice { get; set; }
         public DateTime OrderDate { get; set; }
     }
 }
